pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cairn"
version = "0.1.0"
description = "Multi-agent knowledge sharing system built on the Blackboard Pattern"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "python-multipart>=0.0.9",
    "PyYAML>=6.0.1",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "sse-starlette>=2.1.0",
    "uuid7>=0.1.0",
    "pydantic-settings>=2.3.0",
    "bcrypt>=4.1.0",
    "aiofiles>=23.2.0",
    "jinja2>=3.1.0",
    "httpx-sse>=0.4.0",
    # Phase 3: methodology integration
    "chromadb>=0.5.0",
    # Phase 4: Obsidian vault bridge
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]
nlp = [
    "spacy>=3.0",
]

[project.scripts]
cairn              = "cairn.main:main"
cairn-admin        = "cairn.manage:main"
procedure-validate = "cairn.cli.validate_procedures:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cairn*"]

[tool.setuptools.package-data]
cairn = [
    "db/migrations/*.sql",
    "db/schema/*.sql",
    "ui/*.html",
    "ui/*.css",
    "ui/*.js",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "performance: performance-sensitive tests for sync/search timing budgets",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
